Implementation of starknet state (#146)
* Implementation of starknet state

* Add enum to wrap types

* Finish invoke_constructor function

* Remove update tracker state

* Stylish comments

* Finish declare function implementation, fix clippy errors

* Fix clippy errors

* Simply logic of count_storage_changes function

* fix error in count_storage_changes

* First iteration of tests

* fix declare test in starknet state

* Split internal objects file to simplificate code

* Add deploy test

* Fix format

* Fix clippy

* Fix dependency issues

* Fix broken tests

* Delete unused structure

* Rename apply_specific and apply_concurrent functions

* Fix clippy

* Delete duplicated file

* Fix documentation comments

* Delete duplicated file

* Fix design comments

* Fix test paths

* Fix clippy

* Update CI

+++ src/business_logic/execution/objects.rs
@@ -146,8 +146,8 @@ impl CallInfo {
             for ordered_msg in call.l2_to_l1_messages {
                 let l2tol1msg =
                     L2toL1MessageInfo::new(ordered_msg.clone(), call.caller_address.clone());
-                starknet_events.remove(ordered_msg.order as usize - 1);
-                starknet_events.insert(ordered_msg.order as usize - 1, Some(l2tol1msg));
+                starknet_events.remove(ordered_msg.order - 1);
+                starknet_events.insert(ordered_msg.order - 1, Some(l2tol1msg));
             }
         }
 
@@ -378,7 +378,7 @@ impl TxInfoStruct {
     }
 }
 
-#[derive(Debug, PartialEq, Default)]
+#[derive(Debug, PartialEq, Default, Clone)]
 pub struct TransactionExecutionInfo {
     pub(crate) validate_info: Option<CallInfo>,
     pub(crate) call_info: Option<CallInfo>,
